requirements.txt
mysql-connector-python
python-dotenv
transformers
sentence-transformers
faiss-cpu
PyMuPDF
pandas
numpy
